Disable whitespace-cleanup
whitespace-cleanup has some default settings that exhibit behaviors
that I don't like so this needs to be disabled until I can dedicate a
little bit of time to configuring whitespace-cleanup to not misbehave.

If I can't remember in the future what the problem was, it was this:

There are situations where the default behavior will replace spaces
with a mixture of tabs and spaces, which is never desired.