Move .desktop file to main config Org file
This allows for tangling it directly to the install location, which I
can never remember for .desktop files, for whatever reason. This way
it will be easier to find when I'm looking for it, and also maybe
something else I never have to think about again because it will auto-install.