disable treemacs icons until I can fix them
I'm irritated about this but just turning off the icons and using TUI
mode gets me 90% of the way there and is honestly less to look at.

Still, I'm annoyed at the failure mode of this package and it doesn't
even seem to be caused by some weird interaction with Doom theme stuff
so as far as I can tell it's just an issue in vanilla Treemacs?

IDK.