disable icons in treemacs due to rendering bug
an annoying rendering bug causes directories named bin/ to have a
stupid, ugly icon, that doesn't seem to be dependent on the theme
applied, and that I cannot find a way to control.

Debugging this also made me notice an indentation problem for files
ending in .bash and .sqlite which I had previously applied custom
icons to. I was unable to even get this far with the bin directory due
to limitations with Treemacs's documentation.. it only documents how
to create custom icons for files, and these suffer from the
indentation problem.

Maybe this is due to my config, maybe not, but the bin dir was making
me want to stab my eyes out and at least now, with the icons disabled,
it looks nice again and it is perhaps even nicer than it was to start
with. All I know is that I can see the files in their tree layout and
it doesn't make my eyes bleed anymore. Fine.

Revert this commit if they ever fix the problem upstream. I lost two
and a half hours to trying to fix it here, I'm done. Disabled icons is
fine. Fuck.

🫠