fix stupid gpg ux
using the XDG_SESSION_TYPE env var, tell gpg to prompt for a password
on the cli when applicable instead of just blocking forever because it
can't open a GUI; likewise, avoid the same problem or broken gpg-agent
usage in the GUI / Emacs.

it's so dumb that gpg doesn't just, you know, do this.