split entrypoint and readme.el
the reason to have an init.el file with only the statement that loads
org files is so that we can tangle readme.org at startup into
readme.el and then load it, from inside init.el. If we use init.el to
tangle itself, it won't run the tangled version until next startup.